+++ Mouse.h
@@ -64,7 +64,7 @@ public:
 	bool TriggerMouseButton(MouseButton button) { return nowMouse_.rgbButtons[button] == (0x80) && !(oldMouse_.rgbButtons[button] == (0x80)); }
 
 	// --マウスの位置-- //
-	Vector2 MousePos() { return Vector2(p_.x, p_.y); }
+	Vector2 MousePos() { return Vector2(static_cast<float>(p_.x), static_cast<float>(p_.y)); }
 
 	// --マウスの移動量-- //
 	Vector2 GetMouseVelosity() { return Vector2((float)nowMouse_.lX, (float)nowMouse_.lY); }
